refactor(v3): move discover to new pattern
